engram/storage/pinecone_adapter.py
"""
Pinecone storage adapter — NOT YET IMPLEMENTED.

TODO: Implement vector-based memory using Pinecone.
Similar purpose to ChromaAdapter but for production-scale vector search.

Pinecone docs: https://docs.pinecone.io
Install: pip install pinecone-client

Suggested approach:
- Use Pinecone index per app
- Embed values using an embedding model before upserting
- Metadata filter on key for exact key reads
"""

from __future__ import annotations

from datetime import datetime
from typing import Optional

from engram.models import HistoryEntry, MemoryEntry
from engram.storage.base import StorageAdapter


class PineconeAdapter(StorageAdapter):
    """
    Pinecone storage adapter — NOT YET IMPLEMENTED.

    TODO: Implement vector-based memory using Pinecone.
    Similar purpose to ChromaAdapter but for production-scale vector search.

    Pinecone docs: https://docs.pinecone.io
    Install: pip install pinecone-client

    Suggested approach:
    - Use Pinecone index per app
    - Embed values using an embedding model before upserting
    - Metadata filter on key for exact key reads
    """

    def __init__(self) -> None:
        """
        TODO: Initialize Pinecone client and index.

        Suggested:
            from pinecone import Pinecone
            self._pc = Pinecone(api_key=os.environ["PINECONE_API_KEY"])
            self._index = self._pc.Index("engram")
        """
        raise NotImplementedError("PineconeAdapter is not yet implemented")

    def write(self, entry: MemoryEntry) -> None:
        """
        TODO: Upsert a MemoryEntry into the Pinecone index.

        Suggested: Embed entry.value using an embedding model, then call
        index.upsert(vectors=[{"id": entry.key, "values": embedding, "metadata": {...}}]).
        """
        raise NotImplementedError("PineconeAdapter.write is not yet implemented")

    def read(self, key: str) -> Optional[MemoryEntry]:
        """
        TODO: Fetch a vector by ID from the Pinecone index.

        Suggested: Use index.fetch(ids=[key]) and reconstruct MemoryEntry
        from the metadata.
        """
        raise NotImplementedError("PineconeAdapter.read is not yet implemented")

    def write_history(self, entry: HistoryEntry) -> None:
        """
        TODO: Store a HistoryEntry in Pinecone.

        Suggested: Use a separate namespace "history" with write_id as the vector ID.
        """
        raise NotImplementedError("PineconeAdapter.write_history is not yet implemented")

    def read_history(
        self,
        key: str,
        agent_id: Optional[str] = None,
        since: Optional[datetime] = None,
        until: Optional[datetime] = None,
    ) -> list[HistoryEntry]:
        """
        TODO: Query history entries from Pinecone with metadata filters.

        Suggested: Use index.query() with metadata filters for key, agent_id,
        and timestamp range in the "history" namespace.
        """
        raise NotImplementedError("PineconeAdapter.read_history is not yet implemented")

    def delete(self, key: str) -> None:
        """
        TODO: Delete a vector from Pinecone by ID.

        Suggested: Use index.delete(ids=[key]).
        """
        raise NotImplementedError("PineconeAdapter.delete is not yet implemented")

    def list_keys(self, prefix: Optional[str] = None) -> list[str]:
        """
        TODO: List all vector IDs in the index, optionally filtered by prefix.

        Suggested: Use index.list() with prefix filter if supported,
        otherwise fetch all and filter in Python.
        """
        raise NotImplementedError("PineconeAdapter.list_keys is not yet implemented")

    def ping(self) -> bool:
        """
        TODO: Check that Pinecone is reachable.

        Suggested: Use index.describe_index_stats() and return True if no exception.
        """
        raise NotImplementedError("PineconeAdapter.ping is not yet implemented")
