pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "engram"
version = "0.1.0"
description = "Memory middleware service for multi-agent AI systems"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "redis==5.0.4",
    "httpx==0.27.0",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.0",
    "pytest-asyncio==0.23.6",
    "ruff",
]
vector = [
    "chromadb==0.5.0",
    "pinecone-client==3.2.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.setuptools.packages.find]
include = ["engram*"]
